+++ AHM2_ch09/AHM2_09.4.1__WinBUGS_without_RandomFields.R
@@ -5,10 +5,14 @@
 #
 # Chapter 9 : SPATIAL MODELS OF DISTRIBUTION AND ABUNDANCE
 # ========================================================
-# Code from proofs dated 2020-08-19
 
-if(!requireNamespace("RandomFields"))
-  stop("Package 'RandomFields' is not available.")
+# The code below is modified to run without the RandomFields package; if
+#   RandomFields is available, it will be used by AHMbook, and the results should
+#   be the same.
+# When RandomFields is not available, the 'fields' package is used instead, and
+#   the results will be different.
+if(requireNamespace("RandomFields"))
+  stop("Package 'RandomFields' IS available; this script is not needed.")
 
 # Approximate execution time for this code: 20 mins
 # Run time with the full number of iterations: 3 hrs
@@ -128,26 +132,26 @@ library(R2WinBUGS)
 out2 <- bugs(bdata, inits, params, "CAR.Nmix.txt", n.chains = nc, n.thin = nt,
     n.iter = ni, n.burnin = nb, bugs.directory = bugs.dir)  # ~~~~ for testing
 print(out2$summary[1:10,], 2)
-#               mean      sd    2.5%      25%       50%      75%   97.5% Rhat n.eff
-# mean.lam    3.5569   0.386    2.86    3.289    3.5310    3.799    4.34    1  3000
-# beta0       1.2630   0.109    1.05    1.191    1.2620    1.335    1.47    1  3000
-# beta[1]     1.9796   0.164    1.67    1.865    1.9770    2.091    2.31    1  2100
-# beta[2]    -1.9704   0.136   -2.24   -2.061   -1.9690   -1.878   -1.71    1  2100
-# mean.p      0.5010   0.026    0.45    0.484    0.5013    0.518    0.55    1  2300
-# alpha0      0.0041   0.102   -0.20   -0.065    0.0051    0.071    0.20    1  2300
-# alpha[1]   -0.9065   0.088   -1.08   -0.965   -0.9063   -0.848   -0.73    1  2300
-# alpha[2]   -0.9531   0.066   -1.08   -0.996   -0.9522   -0.909   -0.82    1  3000
-# v.eta       2.3962   0.404    1.69    2.119    2.3670    2.652    3.26    1  2000
-# Ntotal   7235.9227 649.527 6113.90 6771.000 7187.5000 7650.250 8659.17    1  3000
+#              mean      sd    2.5%      25%      50%     75%   97.5% Rhat n.eff
+# mean.lam    4.604   0.407    3.86    4.327    4.584    4.86    5.50  1.0    58
+# beta0       1.523   0.088    1.35    1.465    1.522    1.58    1.71  1.0    60
+# beta[1]     1.979   0.159    1.66    1.869    1.976    2.08    2.29  1.4     9
+# beta[2]    -1.870   0.126   -2.13   -1.956   -1.863   -1.78   -1.64  1.1    19
+# mean.p      0.511   0.022    0.47    0.496    0.511    0.53    0.55  1.0   190
+# alpha0      0.045   0.087   -0.12   -0.016    0.045    0.11    0.21  1.0   200
+# alpha[1]   -0.921   0.072   -1.05   -0.972   -0.921   -0.87   -0.78  1.0    95
+# alpha[2]   -1.065   0.058   -1.18   -1.104   -1.065   -1.02   -0.95  1.0   340
+# v.eta       1.827   0.308    1.32    1.602    1.795    2.02    2.51  1.0    89
+# Ntotal   8272.814 511.584 7335.00 7920.750 8252.000 8599.00 9338.00  1.0   120
 
 # ~~~~ save for comparison with other models ~~~~~~~
-save(out2, file="AHM2_09.4.1_out2.RData")
+save(out2, file="AHM2_09.4.1_out2_without_RandomFields.RData")
 # ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 
 with(dat, cbind(beta0, beta[1], beta[2], alpha0, alpha[1], alpha[2],
     Ntotal = sum(N), summaxC = sum(apply(y,1,max))))
 #      beta0 beta1 beta2 alpha0 alpha1 alpha2 Ntotal summaxC
-# [1,]     2     2    -2      0     -1     -1   7192    4371
+# [1,]     2     2    -2      0     -1     -1   9504    5922
 
 # ~~~ extra code for figure 9.6 ~~~~~
 # Compute average detection probability for each cell
@@ -182,7 +186,7 @@ par(op)
 
 # ~~~ extra code for table of results ~~~~
 # needs out1 object from section
-load("AHM2_09.3_out1.RData")
+load("AHM2_09.3_out1_without_RandomFields.RData")
 truth <- with(dat, c(alpha0 = alpha0, alpha1 = alpha[1], alpha2 = alpha[2],
     beta0 = beta0, beta1 = beta[1], beta2 = beta[2], Ntotal = sum(N)))
 Bayes.est.Nmix0 <- rbind(out1$summary[c('alpha0','alpha[1]', 'alpha[2]',
@@ -191,11 +195,11 @@ Bayes.est.CAR <- rbind(out2$summary[c('alpha0', 'alpha[1]', 'alpha[2]',
     'beta0', 'beta[1]', 'beta[2]', 'Ntotal'), 1:2])
 print(cbind(truth, Bayes.est.Nmix0, Bayes.est.CAR), 2)
 # ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
-#        truth    mean      sd      mean      sd
-# alpha0     0    0.31   0.073    0.0041   0.102
-# alpha1    -1   -0.90   0.067   -0.9065   0.088
-# alpha2    -1   -1.14   0.060   -0.9531   0.066
-# beta0      2    1.29   0.052    1.2630   0.109
-# beta1      2    2.45   0.102    1.9796   0.164
-# beta2     -2   -1.80   0.086   -1.9704   0.136
-# Ntotal  7192 5863.50 238.104 7235.9227 649.527
+#        truth    mean      sd     mean      sd
+# alpha0     0    0.21   0.072    0.045   0.087
+# alpha1    -1   -0.99   0.056   -0.921   0.072
+# alpha2    -1   -1.18   0.053   -1.065   0.058
+# beta0      2    1.81   0.046    1.523   0.088
+# beta1      2    2.06   0.080    1.979   0.159
+# beta2     -2   -1.92   0.079   -1.870   0.126
+# Ntotal  9504 8035.78 311.441 8272.814 511.584