+++ AHM2_ch09/AHM2_09.4.3__WinBUGS.R
@@ -7,6 +7,9 @@
 # ========================================================
 # Code from proofs dated 2020-08-19
 
+if(!requireNamespace("RandomFields"))
+  stop("Package 'RandomFields' is not available.")
+
 # Approximate execution time for this code: 6 mins
 # Run time with the full number of iterations: 1 hr