+++ AHM2_ch10/AHM2_10.5.R
@@ -7,6 +7,7 @@
 # =========================================================
 # Code from proofs dated 2020-08-19
 
+
 # Approximate execution time for this code: 4 mins
 # Run time with full number of iterations: 25 mins